Rename Pages and Redirects folders [branch ch678]
