+++ react/PageList.tsx
@@ -2,9 +2,9 @@ import PropTypes from 'prop-types'
 import React, { PureComponent } from 'react'
 import { Query } from 'react-apollo'
 
-import List from './components/admin/Pages/List'
-import { CategorizedRoutes } from './components/admin/Pages/List/typings'
-import { isStoreRoute } from './components/admin/Pages/utils'
+import List from './components/admin/pages/List'
+import { CategorizedRoutes } from './components/admin/pages/List/typings'
+import { isStoreRoute } from './components/admin/pages/utils'
 import Styles from './components/admin/Styles'
 import Loader from './components/Loader'
 import RoutesQuery from './queries/Routes.graphql'
