+++ react/components/admin/pages/List/Entry.tsx
